fix: recognize that the icon is G2G bigly
The *Add to Home* criteria for Chrome requires that the application have a 192x192 and 512x512 sized icon. I still haven't made the real icon yet, but I believe this is the last thing that's preventing Chrome from displaying the banner. So, let's go ahead and continue to use the default ico one for now and see that we get the app install banner.